refactor: split graphql client into modules
